operator: Add serve command
