+++ tetragonpod/.gitignore
@@ -7,7 +7,6 @@
 *.dylib
 bin
 testbin/*
-Dockerfile.cross
 
 # Test binary, build with `go test -c`
 *.test
